fibonacci

+++ 0x02-functions_nested_loops/102-fibonacci.c
@@ -7,16 +7,16 @@
  */
 int main(void)
 {
-	int n1 = 1;
-	int n2 = 2;
-	int n3;
-	int count;
+	long int n1 = 1;
+	long int n2 = 2;
+	long int n3;
+	long int count;
 
 	printf("1, 2, ");
 	for (count = 3 ; count <= 50 ; count++)
 	{
 		n3 = n1 + n2;
-		printf("%d", n3);
+		printf("%ld", n3);
 		printf(",");
 		n1 = n2;
 		n2 = n3;
